firestore: re-write BackgroundQueue to NOT use the deprecated AsyncTask thread pool (#7376)
This change results in a significant performance improvement of queries
with a large number of documents in their result set.

In my testing, the performance improved by 91% from 3022 ms down to 265
ms on a real device! For some reason, there were _no_ performance gains
on the Android emulator. This must have something to do with how
different Android OS versions manage the threads in the (deprecated)
AsyncTask thread pool.

The test environment was as follows:
- Pixel 7 Pro real device running Android 16.
- Test called collectionRef.whereGreaterThan("foo",
50).get(Source.CACHE) 200 times.
- Calculated the average amount of time it took for the query results to
be received.
- collectionRef is a Firestore collection containing 10,000 documents,
of which 50% match the "whereGreaterThan" filter.
- Local cache had no other documents in it.
- Test app was compiled in "release" mode with
"proguard-android-optimize.txt" r8 configuration.